# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Pivot API
  version: 0.1.0
  summary: The open business intelligence platform for the AI era.
  description: |
    Every non-2xx response uses the same error envelope, carrying a stable
    machine-readable `code`. Codes never change meaning and are never reused,
    so a client may branch on them safely.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  contact:
    name: Pivot
    url: https://github.com/Mmd4LIFE/pivot
servers:
  - url: /
    description: This instance
...
paths:
...
  /api/v1/organization/role-assignments:
    get:
      tags: [authorization]
      operationId: listRoleAssignments
      summary: Who holds which role
      description: Requires `manage_roles`.
      responses:
        "200":
          description: Every role assignment in the caller's organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignmentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    post:
      tags: [authorization]
      operationId: grantRole
      summary: Give a user or group a role
      description: |
        Requires `manage_roles` — which is, deliberately, the most dangerous
        permission in the API: whoever holds it can grant it to themselves.

        Granting the same role twice is a no-op. A subject that does not exist
        in the caller's organization is refused as a validation failure rather
        than stored as a dangling row.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantRoleRequest"
      responses:
        "201":
          description: The role was granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pivot_session
      description: |
        Server-side session. Stateless tokens are deliberately not used for
        browser sessions: they cannot be revoked before expiry, and "remove
        this person's access" has to mean now.

        Set by `POST /api/v1/auth/login` and sent automatically by the
        browser. It is `HttpOnly`, so no client code reads or writes it.
  schemas:
...
    RoleAssignment:
      type: object
      required: [subjectType, subjectId, role]
      properties:
        subjectType:
          type: string
          enum: [user, group]
          description: |
            A `group` assignment is a userset: every member of that group holds
            the role, including members of groups nested beneath it.
        subjectId:
          type: string
          format: uuid
        role:
          type: string
          examples: [admin]
        grantedAt:
          type: string
          format: date-time
    RoleAssignmentList:
      type: object
      required: [assignments]
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/RoleAssignment"
    GrantRoleRequest:
      type: object
      required: [subjectType, subjectId, role]
      properties:
        subjectType:
          type: string
          enum: [user, group]
        subjectId:
          type: string
          format: uuid
        role:
          type: string
          enum: [admin, editor, analyst, viewer]
...
    ErrorDetail:
      type: object
      required: [message]
      properties:
        field:
          type: string
          description: The offending field, for validation failures.
          examples: [email]
        message:
          type: string
          examples: [is required]
    ErrorBody:
      type: object
      required: [code, message, docs]
      properties:
        code:
          type: string
          pattern: "^PIVOT-[A-Z]+-[0-9]{3}$"
          description: |
            A stable, machine-readable identifier. Independent of the HTTP
            status: the status says how the transport should behave, the code
            says what went wrong. Branch on this, never on the message.
          examples: [PIVOT-REQ-003]
        message:
          type: string
          description: A human-readable explanation. May change; do not parse it.
        details:
          type: array
          description: Per-field problems, for validation failures.
          items:
            $ref: "#/components/schemas/ErrorDetail"
        requestId:
          type: string
          description: Correlation ID, echoed in the `X-Request-Id` header.
        docs:
          type: string
          format: uri
          description: Documentation for this error code.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
  responses:
...
    Unauthorized:
      description: Authentication is required, or the tenant could not be resolved
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated, but not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UnprocessableEntity:
      description: The body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: |
        A dependency could not be reached. On a gated endpoint this means the
        authorization backend was unavailable, so the request was denied
        without a decision being made — deliberately distinguishable from 403,
        which means a decision was made and it was no.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
security:
  - sessionCookie: []